[refactor] 회원 탈퇴 닉네임 null 익명화 및 CookieHelper 패키지 이동
- 탈퇴 시 nickname을 '탈퇴한 사용자' 문자열 대신 null로 저장 (DB 완전 익명화)
- UserEntity.toDomain()에 삭제 사용자 변환 가드 추가 (SystemException)
- RefreshTokenCookieHelper를 user.config → user.ui로 이동
- V24 마이그레이션: nickname nullable + 기존 탈퇴 행 null 업데이트 (멱등성 확보)
