+++ backend/src/main/resources/db/migration/V24__nullable_nickname_for_deleted_user.sql
@@ -1,0 +1,2 @@
+ALTER TABLE app_user MODIFY COLUMN nickname VARCHAR(10) NULL;
+UPDATE app_user SET nickname = NULL WHERE deleted_at IS NOT NULL AND nickname IS NOT NULL;
